Refactor associated articles
