+++ src/components/FeedbackForms/AssociatedArticles/index.tsx
@@ -1,0 +1,1 @@
+export * from './RelationTypeDropdown';
