+++ src/components/FeedbackForms/index.ts
@@ -1,2 +1,3 @@
 export * from './MissingRecord';
 export * from './MissingReferences';
+export * from './AssociatedArticles';
